+++ makeRunTest.sh
@@ -24,7 +24,7 @@ fi
 
 make
 
-$CC main.cpp -o main $CFLAGS $LIBS 
+$CC main.cpp testscene.cpp -o main $CFLAGS $LIBS
 
 if [ -f "main" ]; then
 	echo compilation finished!
